Fix blob direct-write obsolete purge race
Recheck blob direct-write ownership before purging blob files so sealed manager-owned files are not deleted before their flush commits. Add obsolete-files coverage for the stale-snapshot purge case.

+++ db/db_impl/db_impl.cc
@@ -2269,6 +2269,10 @@ void DBImpl::BackgroundCallPurge() {
     purge_files_.erase(it);
 
     mutex_.Unlock();
+    if (type == kBlobFile && ShouldKeepBlobFileForPurge(number, fname)) {
+      mutex_.Lock();
+      continue;
+    }
     DeleteObsoleteFileImpl(job_id, fname, dir_to_sync, type, number);
     mutex_.Lock();
   }
